--- a/k6/ENV.csh
+++ b/k6/ENV.csh
@@ -1,6 +1,0 @@
-setenv DEFAULT_DURATION 1s
-setenv DEFAULT_VUS 1
-setenv MIRROR_BASE_URL https://testnet.mirrornode.hedera.com/api/v1
-setenv RELAY_BASE_URL https://testnet.hashio.io/api
-setenv DEFAULT_LIMIT 10
-