--- a/k6/src/scenarios/apis.js
+++ b/k6/src/scenarios/apis.js
@@ -34,7 +34,7 @@ function handleSummary(data) {
 
 function run(testParameters) {
   const scenario = exec.scenario;
-  funcs[scenario.name](testParameters);
+  funcs[scenario.name](testParameters, scenario.iterationInTest);
   scenarioDurationGauge.add(Date.now() - scenario.startTime, {scenario: scenario.name});
 }