--- a/k6/src/scenarios/test/bootstrapEnvParameters.js
+++ b/k6/src/scenarios/test/bootstrapEnvParameters.js
@@ -25,6 +25,8 @@ import {
   setDefaultValuesForEnvParameters
 } from "../../lib/parameters.js";
 
+const scParams = JSON.parse(open("../../prepare/.smartContractParams.json"));
+
 const computeTestParameters = (configuration) =>
   Object.assign({},
     computeLatestContractResultParameters(configuration),
@@ -36,7 +38,7 @@ const setupTestParameters = () => {
   setDefaultValuesForEnvParameters();
   const baseApiUrl = __ENV['MIRROR_BASE_URL'];
   const testParametersMap = computeTestParameters({baseApiUrl: `${baseApiUrl}/api/v1`});
-  return Object.assign(testParametersMap, {
+  return Object.assign(testParametersMap, scParams, {
     MIRROR_BASE_URL: baseApiUrl,
     RELAY_BASE_URL: __ENV['RELAY_BASE_URL'],
     DEFAULT_LIMIT: __ENV.DEFAULT_LIMIT