--- a/k6/src/scenarios/test/common.js
+++ b/k6/src/scenarios/test/common.js
@@ -21,7 +21,12 @@
 const errorField = "error";
 const resultField = "result";
 
+const isDebugMode = __ENV['DEBUG_MODE'] === "true";
+
 function isNonErrorResponse(response) {
+  if(isDebugMode) {
+    console.log(response);
+  }
   //instead of doing multiple type checks,
   //lets just do the normal path and return false,
   //if an exception happens.
@@ -36,12 +41,26 @@ function isNonErrorResponse(response) {
   }
 }
 
+function is400Status(response) {
+  if(isDebugMode) {
+    console.log(response);
+  }
+  try {
+    return response.status === 400;
+  } catch (e) {
+    return false;
+  }
+}
+
 function isErrorResponse(response) {
+  if(isDebugMode) {
+    console.log(response);
+  }
   //instead of doing multiple type checks,
   //lets just do the normal path and return false,
   //if an exception happens.
   try {
-    if (response.status !== 200) {
+    if (response.status === 200) {
       return false;
     }
     const body = JSON.parse(response.body);
@@ -67,4 +86,4 @@ function getPayLoad(methodName, paramInput = []) {
   });
 }
 
-export {isErrorResponse, isNonErrorResponse, httpParams, getPayLoad};
+export {isErrorResponse, isNonErrorResponse, httpParams, getPayLoad, is400Status};