--- a/k6/src/scenarios/test/eth_feeHistory.js
+++ b/k6/src/scenarios/test/eth_feeHistory.js
@@ -28,8 +28,10 @@ const url = __ENV.RELAY_BASE_URL;
 const methodName = 'eth_feeHistory';
 const {options, run} = new TestScenarioBuilder()
   .name(methodName) // use unique scenario name among all tests
-  .request(() => http.post(url, getPayLoad(methodName), httpParams))
+  .request(() => http.post(url, getPayLoad(methodName, ["0x1", "latest", []]), httpParams))
   .check(methodName, (r) => isNonErrorResponse(r))
+  .testDuration("5s")
+  .maxDuration(4500)
   .build();
 
 export {options, run};