--- a/k6/src/scenarios/test/eth_getStorageAt.js
+++ b/k6/src/scenarios/test/eth_getStorageAt.js
@@ -21,16 +21,16 @@
 import http from "k6/http";
 
 import {TestScenarioBuilder} from '../../lib/common.js';
-import {isErrorResponse, httpParams, getPayLoad} from "./common.js";
+import {httpParams, getPayLoad, isNonErrorResponse} from "./common.js";
 import {setupTestParameters} from "./bootstrapEnvParameters.js";
 
 const methodName = 'eth_getStorageAt';
 const {options, run} = new TestScenarioBuilder()
   .name(methodName) // use unique scenario name among all tests
   .request((testParameters) => {
-    return http.post(testParameters.RELAY_BASE_URL, getPayLoad(methodName, [testParameters.DEFAULT_CONTRACT_ADDRESS, "0x0", "latest"]), httpParams);
+    return http.post(testParameters.RELAY_BASE_URL, getPayLoad(methodName, [testParameters.contractAddress, "0x0", "latest"]), httpParams);
   })
-  .check(methodName, (r) => isErrorResponse(r)) // how to scale since dependent on contract
+  .check(methodName, (r) => isNonErrorResponse(r)) // how to scale since dependent on contract
   .build();
 
 export {options, run};