--- a/k6/src/scenarios/test/eth_getTransactionByBlockNumberAndIndex.js
+++ b/k6/src/scenarios/test/eth_getTransactionByBlockNumberAndIndex.js
@@ -28,7 +28,7 @@ const url = __ENV.RELAY_BASE_URL;
 const methodName = 'eth_getTransactionByBlockNumberAndIndex';
 const {options, run} = new TestScenarioBuilder()
   .name(methodName) // use unique scenario name among all tests
-  .request(() => http.post(url, getPayLoad(methodName, ['latest', "0"]), httpParams))
+  .request(() => http.post(url, getPayLoad(methodName, ['latest', "0x0"]), httpParams))
   .check(methodName, (r) => isNonErrorResponse(r))
   .build();