--- a/k6/src/scenarios/test/eth_sign.js
+++ b/k6/src/scenarios/test/eth_sign.js
@@ -21,7 +21,7 @@
 import http from "k6/http";
 
 import {TestScenarioBuilder} from '../../lib/common.js';
-import {isErrorResponse, httpParams, getPayLoad} from "./common.js";
+import {httpParams, getPayLoad, is400Status} from "./common.js";
 
 const url = __ENV.RELAY_BASE_URL;
 
@@ -29,7 +29,7 @@ const methodName = 'eth_sign';
 const {options, run} = new TestScenarioBuilder()
   .name(methodName) // use unique scenario name among all tests
   .request(() => http.post(url, getPayLoad(methodName), httpParams))
-  .check(methodName, (r) => isErrorResponse(r))
+  .check(methodName, (r) => is400Status(r))
   .build();
 
 export {options, run};