… k6/.envexample
… k6/src/prepare/contracts/Greeter.json