**Description**:
- Fixed Typo on README.md
- Included P95 (Percentile 95) on the markdown report at the final step of the performance tests
- Wrapped the project as a node package
- Created npm run task for `prep`, executes prep.js script that bootstrap all the data needed for the tests to run successfully.
- Created npm run task for `k6`, executes the performance tests using k6
- Created npm run task `prep-and-run`, that executes prep and then k6 in sequential order
- Added dynamic payloads for `eth_sendRawTransaction` tests, so it uses pre-generated signedTx using prep.js script, by default only 10tx are pre-generated and signed.
- Added property **FILTER_TEST** for running only 1 or some tests.
- Added property **DEBUG_MODE** for showing logs with the request and response payloads of each test.
- Added property **SIGNED_TXS**, used to determine how many signed trx to generate on prep step.
- Added property **PRIVATE_KEY**, used to run prep.js script.
- Improved output **report.md** with more metrics (median, min, max, p(90) and request)
- Updated the **README.md** file to reflect changes and current instructions and usage.
- The following tests were fixed:
  -  eth_call
  -  eth_coinbase
  -  eth_feeHistory
  -  eth_getBlockByHash
  -  eth_getCode
  -  eth_getStorageAt
  -  eth_getTransactionByBlockNumberAndIndex
  -  eth_getWork
  -  eth_protocolVersion
  -  eth_sendRawTransaction
  -  eth_sendTransaction
  -  eth_sign
  -  eth_signTransaction
  -  eth_submitHashrate

**Related issue(s)**:
N/A

Fixes #867 

**Notes for reviewer**:
<!-- Provide logs, performance numbers or screenshots of the new functionality -->

a. Report MD after adding **P(95)** column:

| Scenario | VUS | Pass% | RPS | Pass RPS | Avg. Req Duration | P(95) Req Duration | Comment |
|----------|-----|-------|-----|----------|-------------------|-------------------------|---------|
| eth_accounts | 1 | 100.00 | 3.95/s | 3.95/s | 109.75ms | 111.10ms | |
| eth_blockNumber | 1 | 100.00 | 3.48/s | 3.48/s | 177.25ms | 208.21ms | |
| eth_call | 1 | 0.00 | 3.32/s | 0.00/s | 0.00ms | 0.00ms | |
| eth_chainId | 1 | 100.00 | 4.68/s | 4.68/s | 103.99ms | 109.89ms | |
| eth_coinbase | 1 | 0.00 | 4.92/s | 0.00/s | 0.00ms | 0.00ms | |
| eth_estimateGas | 1 | 100.00 | 4.96/s | 4.96/s | 98.58ms | 108.36ms | |
| eth_feeHistory | 1 | 0.00 | 4.75/s | 0.00/s | 0.00ms | 0.00ms | |
| eth_gasPrice | 1 | 100.00 | 4.84/s | 4.84/s | 95.98ms | 106.02ms | |
| eth_getBalance | 1 | 100.00 | 3.90/s | 3.90/s | 150.82ms | 164.13ms | |
| eth_getBlockByHash | 1 | 0.00 | 4.86/s | 0.00/s | 0.00ms | 0.00ms | |
| eth_getBlockByNumber | 1 | 0.00 | NaN/s | NaN/s | 0.00ms | 0.00ms | |
| eth_getBlockTransactionCountByHash | 1 | 100.00 | 2.60/s | 2.60/s | 270.25ms | 474.67ms | |
| eth_getBlockTransactionCountByNumber | 1 | 100.00 | 2.46/s | 2.46/s | 301.96ms | 513.89ms | |
| eth_getCode | 1 | 0.00 | 4.78/s | 0.00/s | 0.00ms | 0.00ms | |
| eth_getLogs | 1 | 100.00 | 1.49/s | 1.49/s | 564.78ms | 584.56ms | |
| eth_getStorageAt | 1 | 0.00 | 4.17/s | 0.00/s | 134.45ms | 145.61ms | |
| eth_getTransactionByBlockHashAndIndex | 1 | 100.00 | 3.76/s | 3.76/s | 163.97ms | 198.27ms | |
| eth_getTransactionByBlockNumberAndIndex | 1 | 0.00 | 4.99/s | 0.00/s | 0.00ms | 0.00ms | |
| eth_getTransactionByHash | 1 | 100.00 | 0.28/s | 0.28/s | 3481.97ms | 3481.97ms | |
| eth_getTransactionCount | 1 | 0.00 | NaN/s | NaN/s | 0.00ms | 0.00ms | |
| eth_getTransactionReceipt | 1 | 100.00 | 2.83/s | 2.83/s | 251.45ms | 429.84ms | |
| eth_getUncleByBlockHashAndIndex | 1 | 100.00 | 4.73/s | 4.73/s | 102.73ms | 113.28ms | |
| eth_getUncleByBlockNumberAndIndex | 1 | 100.00 | 4.84/s | 4.84/s | 106.13ms | 119.73ms | |
| eth_getUncleCountByBlockHash | 1 | 100.00 | 4.78/s | 4.78/s | 106.10ms | 126.43ms | |
| eth_getUncleCountByBlockNumber | 1 | 100.00 | 4.78/s | 4.78/s | 102.43ms | 111.53ms | |
| eth_getWork | 1 | 0.00 | 4.84/s | 0.00/s | 0.00ms | 0.00ms | |
| eth_hashrate | 1 | 100.00 | 4.94/s | 4.94/s | 100.47ms | 110.34ms | |
| eth_mining | 1 | 100.00 | 4.61/s | 4.61/s | 104.10ms | 112.61ms | |
| eth_protocolVersion | 1 | 0.00 | 4.90/s | 0.00/s | 0.00ms | 0.00ms | |
| eth_sendRawTransaction | 1 | 0.00 | NaN/s | NaN/s | 0.00ms | 0.00ms | |
| eth_sendTransaction | 1 | 0.00 | 4.83/s | 0.00/s | 0.00ms | 0.00ms | |
| eth_sign | 1 | 0.00 | 4.73/s | 0.00/s | 0.00ms | 0.00ms | |
| eth_signTransaction | 1 | 0.00 | 4.78/s | 0.00/s | 0.00ms | 0.00ms | |
| eth_submitHashrate | 1 | 0.00 | 4.91/s | 0.00/s | 0.00ms | 0.00ms | |
| eth_submitWork | 1 | 100.00 | 4.66/s | 4.66/s | 112.00ms | 142.15ms | |
| eth_syncing | 1 | 100.00 | 4.97/s | 4.97/s | 98.35ms | 104.40ms | |
| net_listening | 1 | 100.00 | 4.87/s | 4.87/s | 101.81ms | 112.05ms | |
| web3_clientVersion | 1 | 100.00 | 4.85/s | 4.85/s | 100.53ms | 108.49ms | |
| web3_client_version | 1 | 100.00 | 4.78/s | 4.78/s | 104.25ms | 113.71ms | |

** I believe that the tests that are showing 0% are the ones that are failing and need fixing.

b. Report AFTER all the changes on this PR:

# K6 Performance Test Results 

JSON-RPC-RELAY URL:  https://previewnet.hashio.io/api

Timestamp: 2023-02-24T22:17:04.051Z 

| Scenario | VUS | Reqs | Pass % | RPS (1/s) | Pass RPS (1/s) | Avg. Req Duration (ms) | Median (ms) | Min (ms) | Max (ms) | P(90) (ms) | P(95) (ms) | Comment |
|----------|-----|------|--------|-----|----------|-------------------|-------|-----|-----|-------|-------|---------|
| eth_accounts | 1 | 5 | 100.00 | 4.53 | 4.53 | 100.38 | 98.79 | 98.76 | 104.00 | 103.03 | 103.51 | |
| eth_blockNumber | 1 | 5 | 100.00 | 4.30 | 4.30 | 134.77 | 132.85 | 131.78 | 143.47 | 139.61 | 141.54 | |
| eth_call | 1 | 3 | 100.00 | 2.87 | 2.87 | 229.14 | 263.49 | 89.26 | 334.66 | 320.42 | 327.54 | |
| eth_chainId | 1 | 5 | 100.00 | 4.99 | 4.99 | 102.37 | 102.72 | 91.55 | 111.75 | 108.49 | 110.12 | |
| eth_coinbase | 1 | 5 | 100.00 | 4.82 | 4.82 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | |
| eth_estimateGas | 1 | 6 | 100.00 | 4.97 | 4.97 | 104.46 | 102.89 | 93.06 | 120.53 | 115.05 | 117.79 | |
| eth_feeHistory | 1 | 11 | 100.00 | 2.58 | 2.58 | 288.02 | 216.88 | 207.32 | 951.04 | 255.90 | 603.47 | |
| eth_gasPrice | 1 | 5 | 100.00 | 4.98 | 4.98 | 103.57 | 102.73 | 97.44 | 109.53 | 108.14 | 108.84 | |
| eth_getBalance | 1 | 5 | 100.00 | 4.14 | 4.14 | 148.01 | 151.74 | 127.70 | 159.03 | 156.97 | 158.00 | |
| eth_getBlockByHash | 1 | 6 | 100.00 | 2.91 | 2.91 | 244.47 | 242.17 | 218.85 | 282.63 | 268.42 | 275.52 | |
| eth_getBlockByNumber | 1 | 4 | 100.00 | 3.69 | 3.69 | 174.80 | 172.27 | 165.09 | 189.59 | 185.05 | 187.32 | |
| eth_getBlockTransactionCountByHash | 1 | 17 | 100.00 | 4.20 | 4.20 | 138.03 | 138.41 | 116.95 | 152.41 | 149.10 | 150.27 | |
| eth_getBlockTransactionCountByNumber | 1 | 7 | 100.00 | 3.49 | 3.49 | 185.96 | 181.94 | 170.07 | 204.35 | 200.33 | 202.34 | |
| eth_getCode | 1 | 3 | 100.00 | 2.82 | 2.82 | 130.03 | 133.64 | 91.96 | 164.50 | 158.33 | 161.41 | |
| eth_getLogs | 1 | 4 | 100.00 | 2.99 | 2.99 | 235.37 | 234.37 | 188.35 | 284.38 | 275.97 | 280.17 | |
| eth_getStorageAt | 1 | 5 | 100.00 | 4.21 | 4.21 | 140.11 | 142.18 | 132.18 | 148.55 | 146.32 | 147.44 | |
| eth_getTransactionByBlockHashAndIndex | 1 | 4 | 100.00 | 3.94 | 3.94 | 159.65 | 162.14 | 135.98 | 178.35 | 175.30 | 176.83 | |
| eth_getTransactionByBlockNumberAndIndex | 1 | 2 | 100.00 | 3.58 | 3.58 | 177.93 | 177.93 | 161.57 | 194.28 | 191.01 | 192.65 | |
| eth_getTransactionByHash | 1 | 4 | 100.00 | 3.31 | 3.31 | 205.91 | 202.36 | 194.31 | 224.59 | 220.23 | 222.41 | |
| eth_getTransactionCount | 1 | 7 | 100.00 | 2.29 | 2.29 | 284.26 | 270.47 | 224.25 | 399.28 | 342.18 | 370.73 | |
| eth_getTransactionReceipt | 1 | 4 | 100.00 | 3.87 | 3.87 | 162.93 | 164.81 | 153.43 | 168.67 | 168.00 | 168.34 | |
| eth_getUncleByBlockHashAndIndex | 1 | 5 | 100.00 | 4.78 | 4.78 | 100.16 | 100.32 | 93.55 | 106.20 | 105.84 | 106.02 | |
| eth_getUncleByBlockNumberAndIndex | 1 | 5 | 100.00 | 4.22 | 4.22 | 136.64 | 107.07 | 93.88 | 262.02 | 204.65 | 233.33 | |
| eth_getUncleCountByBlockHash | 1 | 6 | 100.00 | 5.05 | 5.05 | 98.49 | 97.37 | 93.52 | 105.77 | 104.46 | 105.12 | |
| eth_getUncleCountByBlockNumber | 1 | 5 | 100.00 | 4.94 | 4.94 | 106.07 | 100.91 | 99.06 | 126.72 | 117.39 | 122.05 | |
| eth_getWork | 1 | 5 | 100.00 | 4.93 | 4.93 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | |
| eth_hashrate | 1 | 5 | 100.00 | 4.70 | 4.70 | 107.67 | 111.80 | 94.83 | 117.48 | 115.46 | 116.47 | |
| eth_mining | 1 | 5 | 100.00 | 4.95 | 4.95 | 104.46 | 104.68 | 93.57 | 118.11 | 115.51 | 116.81 | |
| eth_protocolVersion | 1 | 5 | 100.00 | 4.95 | 4.95 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | |
| eth_sendRawTransaction | 1 | 2 | 100.00 | 0.39 | 0.39 | 2462.27 | 2462.27 | 2383.88 | 2540.66 | 2524.98 | 2532.82 | |
| eth_sendTransaction | 1 | 5 | 100.00 | 4.85 | 4.85 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | |
| eth_sign | 1 | 3 | 100.00 | 2.33 | 2.33 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | |
| eth_signTransaction | 1 | 6 | 100.00 | 5.10 | 5.10 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | |
| eth_submitHashrate | 1 | 6 | 100.00 | 5.05 | 5.05 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | |
| eth_submitWork | 1 | 5 | 100.00 | 4.84 | 4.84 | 107.37 | 102.09 | 98.70 | 132.91 | 120.87 | 126.89 | |
| eth_syncing | 1 | 6 | 100.00 | 5.04 | 5.04 | 99.14 | 99.96 | 90.02 | 109.07 | 105.16 | 107.12 | |
| net_listening | 1 | 5 | 100.00 | 4.80 | 4.80 | 106.01 | 104.28 | 94.85 | 122.19 | 117.98 | 120.09 | |
| web3_clientVersion | 1 | 6 | 100.00 | 5.09 | 5.09 | 100.09 | 101.36 | 91.31 | 105.01 | 104.75 | 104.88 | |
| web3_client_version | 1 | 5 | 100.00 | 5.00 | 5.00 | 104.68 | 100.93 | 96.54 | 118.14 | 115.36 | 116.75 | |



**Checklist**

- [x] Documented (Code comments, README, etc.)
- [x] Tested (unit, integration, etc.) **Only manual testing**