@AlfredoG87 2023-02-21
> Do we care about other P values (P50, P99)?

@lukelee-sl, we are already persisting the Average or (P50).

I also thought on including P(99) but was not sure if we wanted that information, I could add it if you think we should. personally I was not sure it would be valuable.