@Nana-EC 2023-02-22
@AlfredoG87 is this the version that doesn't add setup logic to populate contracts etc so that `eth_call` and `eth_sendRawTransaction` work?