@AlfredoG87 2023-02-22
> @AlfredoG87 is this the version that doesn't add setup logic to populate contracts etc so that `eth_call` and `eth_sendRawTransaction` work?

@Nana-EC, That is correct. running them against previewnet or testnet, there is no problem, only issue is with local-node.